Move query_sbwt to index.rs
